Now pulls values into an array, sorts those, then re-creates the table based off that - result is several orders of magnitude faster.
For example, with a dataset I was testing, at about 250 rows, it was taking >6s to sort! Now the entire dataset of >4500 rows takes <1s to sort.

I'm sure it can still be improved more, though of course we're relying on whichever browser's underlying Array.prototype.sort() method, which is QuickSort at least in Chrome (and likely elsewhere)

Likely the default comparison function can be further optimized to reduce branching and checks within the actual comparison, but the performance is still much better.

Also, now there are some options to allow default sorting, selecting the class for auto-sorting and deciding whether to sort by the value of the first contained element.